back/pyproject.toml
[tool.poetry]
name = "fastapi_sqlmodel"
version = "0.1.1"
description = ""
authors = ["hotratx <mgviero@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10.2"
uvicorn = "^0.15.0"
asyncpg = "^0.24.0"
fastapi = {extras = ["all"], version = "^0.70.0"}
alembic = "^1.7.4"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
broadcaster = {extras = ["redis"], version = "^0.2.0"}
aiosmtplib = "^1.1.6"
aioredis = "^2.0.1"
dependency-injector = "^4.38.0"
email-validator = "^1.1.3"
sqlmodel = "^0.0.6"
python-dotenv = "^0.20.0"
Jinja2 = "^3.1.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
flake8 = "^4.0.1"
mypy = "^0.910"
httpx = "^0.21.1"
asgi-lifespan = "^1.0.1"
pytest-asyncio = "^0.16.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
